+++ test/demo_defkey.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright (c) 2002-2008,2012 Free Software Foundation, Inc.              *
+ * Copyright (c) 2002-2012,2013 Free Software Foundation, Inc.              *
  *                                                                          *
  * Permission is hereby granted, free of charge, to any person obtaining a  *
  * copy of this software and associated documentation files (the            *
@@ -26,7 +26,7 @@
  * authorization.                                                           *
  ****************************************************************************/
 /*
- * $Id: demo_defkey.c,v 1.21 2012/12/29 22:55:39 tom Exp $
+ * $Id: demo_defkey.c,v 1.22 2013/09/28 22:02:17 tom Exp $
  *
  * Demonstrate the define_key() function.
  * Thomas Dickey - 2002/11/23
@@ -116,7 +116,7 @@ visible(const char *string)
 		result = typeCalloc(char, need);
 	}
     } else {
-	result = typeCalloc(char, 1);
+	result = typeCalloc(char, (size_t) 1);
     }
     return result;
 }
@@ -186,9 +186,9 @@ duplicate(WINDOW *win, NCURSES_CONST char *name, int code)
 	const char *prefix = 0;
 	char temp[BUFSIZ];
 
-	if (!strncmp(value, "\033[", 2)) {
+	if (!strncmp(value, "\033[", (size_t) 2)) {
 	    prefix = "\033O";
-	} else if (!strncmp(value, "\033O", 2)) {
+	} else if (!strncmp(value, "\033O", (size_t) 2)) {
 	    prefix = "\033[";
 	}
 	if (prefix != 0) {
