+++ test/edit_field.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright (c) 2003-2008,2011 Free Software Foundation, Inc.              *
+ * Copyright (c) 2003-2011,2013 Free Software Foundation, Inc.              *
  *                                                                          *
  * Permission is hereby granted, free of charge, to any person obtaining a  *
  * copy of this software and associated documentation files (the            *
@@ -26,7 +26,7 @@
  * authorization.                                                           *
  ****************************************************************************/
 /*
- * $Id: edit_field.c,v 1.21 2013/06/08 21:38:12 tom Exp $
+ * $Id: edit_field.c,v 1.22 2013/09/28 22:02:17 tom Exp $
  *
  * A wrapper for form_driver() which keeps track of the user's editing changes
  * for each field, and makes the resulting length available as a
@@ -329,7 +329,7 @@ init_edit_field(FIELD * f, char *value)
     if (ptr == 0) {
 	int rows, cols, frow, fcol, nrow, nbuf;
 
-	ptr = typeCalloc(FieldAttrs, 1);
+	ptr = typeCalloc(FieldAttrs, (size_t) 1);
 	ptr->background = field_back(f);
 	if (field_info(f, &rows, &cols, &frow, &fcol, &nrow, &nbuf) == E_OK) {
 	    ptr->row_count = nrow;
